Rename test files and add build tags for integration
